♻️ Refactor: 솔루션 준비/진행 화면 합병
- 관련: #23
