--- app/src/main/java/com/example/saybettereducator/ui/intent/VideoCallIntent.kt
@@ -1,9 +1,0 @@
-package com.example.saybettereducator.ui.intent
-
-import com.example.saybettereducator.ui.sideeffect.VideoCallSideEffect
-
-sealed class VideoCallIntent {
-    object OnRemoteViewReady: VideoCallIntent()
-    object SetupView: VideoCallIntent()
-    data class SendRTCMessage(val msg: String) : VideoCallIntent()
-}
