+++ app/src/main/java/com/example/saybettereducator/ui/sideeffect/SessionSideEffect.kt
@@ -1,0 +1,5 @@
+package com.example.saybettereducator.ui.sideeffect
+
+sealed class SessionSideEffect {
+    object PeerConnectionSuccess : SessionSideEffect()
+}
