--- app/src/main/java/com/example/saybettereducator/ui/sideeffect/VideoCallSideEffect.kt
@@ -1,5 +1,0 @@
-package com.example.saybettereducator.ui.sideeffect
-
-sealed class VideoCallSideEffect {
-    object PeerConnectionSuccess : VideoCallSideEffect()
-}
